DS-351: Documentation and create single story file
